tests/wpt/web-platform-tests/css/css-backgrounds/parsing/border-image-width-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Backgrounds and Borders Module Level 3: parsing border-image-width with valid values</title>
<link rel="author" title="Eric Willigers" href="mailto:ericwilligers@chromium.org">
<link rel="help" href="https://drafts.csswg.org/css-backgrounds/#border-image-width">
<meta name="assert" content="border-image-width supports the full grammar '[ <length-percentage> | <number> | auto ]{1,4}'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("border-image-width", "1px");
test_valid_value("border-image-width", "2%");
test_valid_value("border-image-width", "3");
test_valid_value("border-image-width", "auto");

test_valid_value("border-image-width", "1px 2% 3 auto");
</script>
</body>
</html>
